move production code into src folder to diff from test, temporarily render the app unusable
