+++ src/application/opencv/Image.py
