+++ src/infrastructure/openCV/draw/show_frames.py
