+++ src/model/keypoint_classifier/keypoint_classifier.hdf5
