+++ src/model/keypoint_classifier/keypoint_classifier.tflite
