+++ src/model/keypoint_classifier/keypoint_classifier_label.csv
